Add default values to interface SelectionState
This commit adds default values to the SelectionState interface.

This is useful to avoid unnecessary empty default implementations, while still allowing the method to be overriden.

The tests added can infer correctly that the date selected is false, but this author couldn't think of a better test to "thid method shouldn't do anything", so I've added a check to ensure no exceptions are thrown.

Closes #58